external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/rendering/gl-scissor-test-expected.txt

Check if glScissor setting works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Canvas.getContext
PASS context exists

PASS pixel at 0, 0 is expected value
PASS pixel at 1, 0 is expected value
PASS pixel at 0, 1 is expected value
PASS pixel at 1, 1 is expected value

PASS successfullyParsed is true

TEST COMPLETE
